digital illustrator board/requirements.txt.txt
opencv-python
mediapipe
customtkinter
pillow
numpy